Properly scale and position glyphs

+++ src/main/kotlin/org/tobi29/scapes/engine/gui/GlyphRenderer.kt
@@ -28,8 +28,8 @@ interface GlyphRenderer {
 
     fun pageCode(character: Char): Int
 
-    class GlyphPage(val width: FloatArray,
+    class GlyphPage(val width: IntArray,
                     val size: Int,
                     val tiles: Int,
-                    val tileSize: Float)
+                    val tileSize: Double)
 }
